Extract Advise workspace client boundary
